common: move download functions to download.(cpp|h)
